Significantly reduce Output HUD's GPU and CPU usage
There are many changes in this commit, but the main contributors to use
less GPU resources are: scissoring and disabling of alpha blending.

The neon effect allows an overlay over an whole output, but in practice,
the window list only takes about half the output's height (typically
less, at least how I use workspaces) and half the output's width. This
means we render a lot of unnecessary pixels. With a 4k display attached,
this means, due to neon's blurring, processing 160 million pixels. Not
feasible. So, instead, limit neon's rendering to the actual overlay size
(i.e. the window list) via scissoring.

Additionaly, large parts of the neon effect don't require alpha
blending. Hence, disable it there.

And finally, there are some more optimizations:

Ḿove rendering of rectangles to OpenGL. Doing it in Cairo means dealing
with pixels three times: once for clearing (or initializing) the cairo
surface, once for drawing the rectangle and a third time for copying the
cairo surface to an OpenGL texture. Now, the rectangles are drawn
directly with OpenGL and only texts are drawn with Cairo. As another
optimization, the rendered texts (i.e. the window titles and workspace
names) are cached in an LRU cache.

Showing the default pointer also means updating an OpenGL texture. So we
now check whether it's really required to show the default pointer (i.e.
didn't we show the pointer before already). This should only be a minor
optimization really, at least I personally basically never view empty
workspaces.